requirements.txt
babel==2.17.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
courlan==1.3.2
dateparser==1.2.1
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
htmldate==1.9.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
jusText==3.0.2
lxml==5.3.2
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
numpy==2.2.4
packaging==24.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
SQLAlchemy==2.0.40
threadpoolctl==3.6.0
tld==0.13
trafilatura==2.0.0
typing_extensions==4.13.1
tzlocal==5.3.1
urllib3==2.3.0
Werkzeug==3.1.3
ci-cloudconnector
